Stock Service Completed

+++ Order.API/Models/OrderItem.cs
@@ -6,7 +6,7 @@ public class OrderItem
 {
     public int Id { get; set; }
     
-    public decimal ProductId { get; set; }
+    public int ProductId { get; set; }
 
     [Column(TypeName = "decimal(18,2)")] public decimal Price { get; set; }
